Score Script
Pushing before I really start to break things--
A couple things of note here: A new script attached to the top level canvas called ScoreScript has been made-- this is for score calculations, and for holding all the player's progress through a round.
Also, right now if you click too quickly through the fledgling screens the game will crash. This is problematic.
Also, I renamed fledgling and mister to "apprentice" and "chef" respectively, because it was way easier to make work in code.